requirements.txt
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
mysql-connector-python>=8.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
reportlab>=4.0.0
pydantic==2.6.0
email-validator>=2.0.0
python-multipart==0.0.6